# Conflicts:
#	src/main/java/com/spring/boot/controller/OrderController.java
#	src/main/java/com/spring/boot/dao/PayRepository.java
#	src/main/java/com/spring/boot/model/Product.java
#	src/main/java/com/spring/boot/model/SiteUser.java
#	src/main/java/com/spring/boot/service/PayService.java
#	src/main/java/com/spring/boot/service/ProductService.java

브랜치통합

+++ src/main/java/com/spring/boot/controller/AnswerController.java
@@ -65,7 +65,7 @@ public class AnswerController {
 		
 		
 		Product product =question.getProduct();
-		Integer pno=product.getId();
+		Long pno=product.getId();
 		
 		//다시 상품페이지detail 로 돌아오기
 		return String.format("redirect:/product/detail/%s", pno);
